cleanup some more stuff
